+++ changelog.md
@@ -21,12 +21,15 @@
     - Fixed bold black and nested bold colors when generating color previews
     - Fixed virtual area 3d map adding invalid exits when created from wizard
     - Fixed resize dialog when changing level anchor
+    - Fixed when closing cancel from save not preventing close
     - Virtual area editor: 
       - Change code in generation of external room to use BASEROOM define
       - Fixed extra map codes being added to rooms when building raw map files
       - Fixed terrain/state files not having correct sizes when maps resized
       - Fixed issue when opening map it would lose sub file data
       - Fixed drag and drop of text when dealing with raw text editors
+      - Fixed changing external exits not correctly updating visual map
+      - Fixed adding new external exit breaking last added external exit
 - **Changed:**
   - Update electron 31.1.0 to 32.0.1
   - Update better-sqlite3 11.1.1 to 11.2.1